+++ ast/meta.go
@@ -20,7 +20,7 @@ import (
 	"fmt"
 	"strings"
 
-	"github.com/XiaoMi/soar/common"
+	"github.com/sjatsh/soar/common"
 	"vitess.io/vitess/go/vt/sqlparser"
 )
 
@@ -297,8 +297,11 @@ func FindEQColsInWhere(node sqlparser.SQLNode) []*common.Column {
 				}
 
 			default:
-				if _, ok := eqOperators[node.Operator]; ok {
-					newCols = FindColumn(node)
+				switch node.Left.(type) {
+				case *sqlparser.ColName:
+					if _, ok := eqOperators[node.Operator]; ok {
+						newCols = FindColumn(node)
+					}
 				}
 			}