+++ ast/tidb_test.go
@@ -20,7 +20,7 @@ import (
 	"fmt"
 	"testing"
 
-	"github.com/XiaoMi/soar/common"
+	"github.com/sjatsh/soar/common"
 )
 
 func TestPrintPrettyStmtNode(t *testing.T) {
@@ -53,6 +53,64 @@ func TestStmtNode2JSON(t *testing.T) {
 	}, t.Name(), update)
 	if nil != err {
 		t.Fatal(err)
+	}
+	common.Log.Debug("Exiting function: %s", common.GetFunctionName())
+}
+
+func TestSchemaMetaInfo(t *testing.T) {
+	common.Log.Debug("Entering function: %s", common.GetFunctionName())
+	sqls := []string{
+		"use world_x;",
+		"select 1;",
+		"syntax error case",
+		"select * from ta join tb using (id)",
+		"select * from ta, tb limit 1",
+		"drop table tb",
+		"drop table db.tb",
+		"drop database db",
+		"create database db",
+		"create index idx_col on tbl (col)",
+		"DROP INDEX idx_col on tbl",
+	}
+	// fmt.Println(sqls[len(sqls)-1])
+	// fmt.Println(SchemaMetaInfo(sqls[len(sqls)-1], "sakila"))
+	// return
+	err := common.GoldenDiff(func() {
+		for _, sql := range append(sqls, common.TestSQLs...) {
+			fmt.Println(sql)
+			fmt.Println(SchemaMetaInfo(sql, "sakila"))
+		}
+	}, t.Name(), update)
+	if nil != err {
+		t.Fatal(err)
+	}
+	common.Log.Debug("Exiting function: %s", common.GetFunctionName())
+}
+
+func TestRemoveIncompatibleWords(t *testing.T) {
+	common.Log.Debug("Entering function: %s", common.GetFunctionName())
+	sqls := [][]string{
+		{
+			`CREATE TEMPORARY TABLE IF NOT EXISTS t_film AS (SELECT * FROM film)`,
+			`CREATE CONSTRAINT col_fk FOREIGN KEY (col) REFERENCES tb (id) ON UPDATE CASCADE`,
+			"CREATE FULLTEXT KEY col_fk (col) /*!50100 WITH PARSER `ngram` */",
+			`CREATE /*!50100 PARTITION BY LIST (col)`,
+			`CREATE col varchar(10) CHARACTER SET gbk DEFAULT NULL`,
+		},
+		{
+			`CREATE TABLE IF NOT EXISTS t_film AS (SELECT * FROM film)`,
+			`CREATE CONSTRAINT col_fk FOREIGN KEY (col) REFERENCES tb (id)`,
+			"CREATE FULLTEXT KEY col_fk (col) /* 50100 WITH PARSER `ngram` */",
+			`CREATE /* 50100 PARTITION BY LIST (col)`,
+			`CREATE col varchar(10) DEFAULT NULL`,
+		},
+	}
+	for k, sql := range sqls[0] {
+		sql = removeIncompatibleWords(sql)
+		if sqls[1][k] != sql {
+			fmt.Println(sql)
+			t.Fatal(sql)
+		}
 	}
 	common.Log.Debug("Exiting function: %s", common.GetFunctionName())
 }