+++ ast/token.go
@@ -342,10 +342,6 @@ var TokenString = map[int]string{
 	sqlparser.ZEROFILL:                "zerofill",
 	sqlparser.DATABASES:               "databases",
 	sqlparser.TABLES:                  "tables",
-	sqlparser.VITESS_KEYSPACES:        "vitess_keyspaces",
-	sqlparser.VITESS_SHARDS:           "vitess_shards",
-	sqlparser.VITESS_TABLETS:          "vitess_tablets",
-	sqlparser.VSCHEMA_TABLES:          "vschema_tables",
 	sqlparser.NAMES:                   "names",
 	sqlparser.CHARSET:                 "charset",
 	sqlparser.GLOBAL:                  "global",
@@ -366,6 +362,8 @@ var TokenString = map[int]string{
 	sqlparser.SUBSTRING:               "substring",
 	sqlparser.GROUP_CONCAT:            "group_concat",
 	sqlparser.SEPARATOR:               "separator",
+	sqlparser.VSCHEMA:                 "vschema",
+	sqlparser.SEQUENCE:                "sequence",
 	sqlparser.MATCH:                   "match",
 	sqlparser.AGAINST:                 "against",
 	sqlparser.BOOLEAN:                 "boolean",
@@ -446,6 +444,7 @@ var mySQLKeywords = map[string]string{
 	"geometry":           "GEOMETRY",
 	"geometrycollection": "GEOMETRYCOLLECTION",
 	"global":             "GLOBAL",
+	"grant":              "GRANT",
 	"group":              "GROUP",
 	"group_concat":       "GROUP_CONCAT",
 	"having":             "HAVING",
@@ -512,6 +511,7 @@ var mySQLKeywords = map[string]string{
 	"reorganize":         "REORGANIZE",
 	"repair":             "REPAIR",
 	"replace":            "REPLACE",
+	"revoke":             "REVOKE",
 	"right":              "RIGHT",
 	"rlike":              "REGEXP",
 	"rollback":           "ROLLBACK",
@@ -926,13 +926,18 @@ func SplitStatement(buf []byte, delimiter []byte) (string, string, []byte) {
 		}
 
 		// quoted string
-		if b == '`' || b == '\'' || b == '"' {
+		switch b {
+		case '`', '\'', '"':
 			if i > 1 && buf[i-1] != '\\' {
 				if quoted && b == quoteRune {
 					quoted = false
+					quoteRune = 0
 				} else {
-					quoted = true
-					quoteRune = b
+					// check if first time found quote
+					if quoteRune == 0 {
+						quoted = true
+						quoteRune = b
+					}
 				}
 			}
 		}
@@ -985,4 +990,18 @@ func NewLines(buf []byte) int {
 		}
 	}
 	return newLines
+}
+
+// QueryType get query type such as SELECT/INSERT/DELETE/CREATE/ALTER
+func QueryType(sql string) string {
+	tokens := Tokenize(sql)
+	for _, token := range tokens {
+		// use strings.Fields for 'ALTER TABLE' token split
+		for _, tk := range strings.Fields(strings.TrimSpace(token.Val)) {
+			if val, ok := mySQLKeywords[strings.ToLower(tk)]; ok {
+				return val
+			}
+		}
+	}
+	return "UNKNOWN"
 }